requirements.txt
blinker==1.9.0
# Core Flask dependencies
Flask==3.1.1
Flask-WTF==1.2.2
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.1
WTForms==3.2.1

# Database & Auth
supabase==2.16.0
gotrue==2.12.2
postgrest==1.1.1
realtime==2.5.3

# Google AI
google-generativeai==0.8.5
google-auth==2.40.3
google-api-core==2.25.1
protobuf==4.25.8

# HTTP/API
httpx==0.28.1
httpcore==1.0.9
requests==2.32.4

# Utilities
python-dotenv==1.1.0
pydantic==2.11.7
pydantic_core==2.33.2
anyio==4.9.0

# Production
gunicorn==21.2.0

# Development
pytest==8.4.1
pytest-mock==3.14.1
